fix(plc4j/opcua): Fix null pointer exception in OPC UA Subscription Handle.
The 'sampling' mode in case of OPC-UA means that no actual events are generated to the client.
This is a sort of 'suspended' subscription which will deliver its events once monitoring item is switched back to reporting mode.
In essence all UA subscriptions must be in reporting mode to generate data for client.

Closes #2024.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>